refactor: 使わなくなったクラスを削除、構成を統一するためにフォルダ構成を整え・移動。
