+++ src/DIContainerBuilder.ts
@@ -7,7 +7,7 @@ import UserService from './application/service/UserService';
 
 // DI auto register controllers.
 import './presentation/apiv1/user/UserController';
-import './presentation/apiv1/PhotoListController';
+import './presentation/apiv1/photo/PhotoListController';
 
 export default class DIContainerBuilder {
   constructor(private readonly settings: Settings) {}
@@ -24,12 +24,6 @@ export default class DIContainerBuilder {
 
     container.bind<UserRepository>('UserRepository').to(UserDatasource);
     container.bind<UserService>('UserService').to(UserService);
-
-    // container.bind<UserAddApi>('UserAddApi').to(UserAddApi);
-    // container.bind<UserApi>('UserApi').to(UserApi);
-    // container.bind<UsersApi>('UsersApi').to(UsersApi);
-
-    // container.bind<ApiV1RouterWrapper>('ApiV1RouterWrapper').to(ApiV1RouterWrapper);
   }
 
   private createConnectedDbPool(): Pool {
