--- src/presentation/HttpMethod.ts
@@ -1,6 +1,0 @@
-enum HttpMethod {
-    Get,
-    Post
-}
-
-export default HttpMethod;
