+++ src/presentation/apiv1/photo/PhotoListController.ts
